# lab07/Source/src/main/resources/openapi.yaml
{"openapi":"3.0.1","info":{"title":"InternetProvider - OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/installations/{id}":{"put":{"tags":["installation-controller"],"operationId":"updateInstallation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Installation"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Installation"}}}}}}},"/api/clients/{id}":{"get":{"tags":["client-controller"],"operationId":"getClientById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Client"}}}}}},"put":{"tags":["client-controller"],"operationId":"updateClient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Client"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Client"}}}}}},"delete":{"tags":["client-controller"],"operationId":"deleteClient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/installations":{"post":{"tags":["installation-controller"],"operationId":"createInstallation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Installation"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Installation"}}}}}}},"/api/clients":{"get":{"tags":["client-controller"],"operationId":"getAllClients","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Client"}}}}}}},"post":{"tags":["client-controller"],"operationId":"addClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Client"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Client"}}}}}}},"/installations/{installationId}":{"get":{"tags":["installation-controller"],"operationId":"getInstallationById","parameters":[{"name":"installationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Installation"}}}}}},"delete":{"tags":["installation-controller"],"operationId":"deleteInstallationById","parameters":[{"name":"installationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/greeting":{"get":{"tags":["my-controller"],"operationId":"greeting","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string","default":"World"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Charges":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"paymentDueDate":{"type":"string","format":"date"},"amountToPay":{"type":"number"},"installation":{"$ref":"#/components/schemas/Installation"}}},"Installation":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"address":{"type":"string"},"routerNumber":{"type":"integer","format":"int64"},"serviceType":{"type":"string"},"chargesList":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Charges"}},"paymentList":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Payment"}}}},"Payment":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"paymentDate":{"type":"string","format":"date"},"amountPaid":{"type":"number"},"installation":{"$ref":"#/components/schemas/Installation"}}},"Client":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"surname":{"type":"string"}}}}}}